Loader support, framework, shell extension, and example
This adds the framework support to build executables for Nautilus
and a loader within Nautilus that can dynamically load, execute, and
unload these executables.   An executable runs as part of the kernel
and can do anything.  A (limited) set of kernel functions are exposed
to it via src/nautilus/loader.c and include/nautilus/nautilus_exe.h.

An executable is currently restricted.  It's essentially a cross between
a Linux kernel module, a DOS COM file, and a dark shadow capsule.